src/requirements.txt
defusedxml==0.4.1
dj-stripe==0.8.0
Django==1.10.1
django-allauth==0.27.0
django-braces==1.9.0
django-model-utils==2.6
jsonfield==1.0.3
oauthlib==2.0.0
python3-openid==3.0.10
pytz==2016.6.1
requests==2.11.1
requests-oauthlib==0.7.0
stripe==1.38.0
gunicorn
djangorestframework==3.6.3
django-push-notifications==1.5.0
redis==2.10.5
mysqlclient==1.3.10